+++ crates/coeus-hephaestus/src/reduction.rs
@@ -48,7 +48,7 @@ pub enum ScanOperation {
 }
 
 /// Dispatches Coeus reduction requests through one Hephaestus axis seam.
-pub trait AxisReductionDispatch<D: ComputeDevice, T: bytemuck::Pod> {
+pub trait AxisReductionDispatch<D: ComputeDevice, T: bytemuck::Pod + eunomia::Pod> {
     /// Execute one Coeus reduction operation.
     fn reduce(
         device: &D,
@@ -63,6 +63,7 @@ impl<D, T, R> AxisReductionDispatch<D, T> for R
 where
     D: ComputeDevice,
     T: bytemuck::Pod
+        + eunomia::Pod
         + hephaestus_core::DialectScalar<R::Dialect>
         + OpIdentity<SumOp>
         + OpIdentity<hephaestus_core::ProdOp>
@@ -101,7 +102,7 @@ where
 }
 
 /// Dispatches Coeus scans through one Hephaestus scan seam.
-pub trait ScanDispatch<D: ComputeDevice, T: bytemuck::Pod> {
+pub trait ScanDispatch<D: ComputeDevice, T: bytemuck::Pod + eunomia::Pod> {
     /// Execute one Coeus scan operation.
     fn scan(
         device: &D,
@@ -117,6 +118,7 @@ impl<D, T, S> ScanDispatch<D, T> for S
 where
     D: ComputeDevice,
     T: bytemuck::Pod
+        + eunomia::Pod
         + hephaestus_core::DialectScalar<S::Dialect>
         + OpIdentity<hephaestus_core::CumSumOp>
         + OpIdentity<hephaestus_core::CumProdOp>
@@ -208,7 +210,7 @@ pub struct HephaestusBackend<P>(std::marker::PhantomData<P>);
 struct ScanRequest<'a, P, T>
 where
     P: HephaestusProvider,
-    T: bytemuck::Pod,
+    T: bytemuck::Pod + eunomia::Pod,
 {
     input: &'a HephaestusStorage<P, T>,
     input_layout: &'a Layout,